Adds mood history tracker
Initializes an empty list and treats it like a queue when adding and removing (FIFO), starts removing when a new string would make the list length greater than 3.
Creates a stateless widget called MoodHistory to display the list, uses conditional rendering to show a text when the list is empty.